Extract core parts of polkadot.
