+++ primitives/chains/polkadot/Cargo.toml
@@ -9,28 +9,19 @@ license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
 [dependencies]
 
 # Bridge Dependencies
-
 bp-message-lane = { path = "../../message-lane", default-features = false }
-bp-runtime = { path = "../../runtime", default-features = false }
+bp-polkadot-core = { path = "../../polkadot-core", default-features = false }
 
 # Substrate Based Dependencies
 
-frame-support = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-frame-system = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
 sp-api = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "master", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "master", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate.git", branch = "master", default-features = false }
 
 [features]
 default = ["std"]
 std = [
 	"bp-message-lane/std",
-	"bp-runtime/std",
-	"frame-support/std",
-	"frame-system/std",
+	"bp-polkadot-core/std",
 	"sp-api/std",
-	"sp-core/std",
-	"sp-runtime/std",
 	"sp-std/std",
 ]
