+++ primitives/polkadot-core/src/lib.rs
@@ -1,4 +1,4 @@
-// Copyright 2019-2020 Parity Technologies (UK) Ltd.
+// Copyright 2019-2021 Parity Technologies (UK) Ltd.
 // This file is part of Parity Bridges Common.
 
 // Parity Bridges Common is free software: you can redistribute it and/or modify
@@ -15,19 +15,84 @@
 // along with Parity Bridges Common.  If not, see <http://www.gnu.org/licenses/>.
 
 #![cfg_attr(not(feature = "std"), no_std)]
-// RuntimeApi generated functions
-#![allow(clippy::too_many_arguments)]
-// Runtime-generated DecodeLimit::decode_all_with_depth_limit
-#![allow(clippy::unnecessary_mut_passed)]
 
-use bp_message_lane::{LaneId, Weight, MessageNonce};
-use sp_std::prelude::*;
+use bp_message_lane::MessageNonce;
+use bp_runtime::Chain;
+use frame_support::RuntimeDebug;
+use sp_core::Hasher as HasherT;
+use sp_runtime::{
+	generic,
+	traits::{BlakeTwo256, IdentifyAccount, Verify},
+	MultiSignature, OpaqueExtrinsic as UncheckedExtrinsic,
+};
 
-pub use bp_polkadot_core::*;
+// TODO: may need to be updated after https://github.com/paritytech/parity-bridges-common/issues/78
+/// Maximal number of messages in single delivery transaction.
+pub const MAX_MESSAGES_IN_DELIVERY_TRANSACTION: MessageNonce = 128;
 
-/// Polkadot Chain
-pub type Polkadot = PolkadotLike;
+/// Maximal number of unrewarded relayer entries at inbound lane.
+pub const MAX_UNREWARDED_RELAYER_ENTRIES_AT_INBOUND_LANE: MessageNonce = 128;
 
+// TODO: should be selected keeping in mind:
+// finality delay on both chains + reward payout cost + messages throughput.
+/// Maximal number of unconfirmed messages at inbound lane.
+pub const MAX_UNCONFIRMED_MESSAGES_AT_INBOUND_LANE: MessageNonce = 8192;
+
+/// Block number type used in Polkadot-like chains.
+pub type BlockNumber = u32;
+
+/// Hash type used in Polkadot-like chains.
+pub type Hash = <BlakeTwo256 as HasherT>::Out;
+
+/// The type of an object that can produce hashes on Polkadot-like chains.
+pub type Hasher = BlakeTwo256;
+
+/// The header type used by Polkadot-like chains.
+pub type Header = generic::Header<BlockNumber, Hasher>;
+
+/// Signature type used by Polkadot-like chains.
+pub type Signature = MultiSignature;
+
+/// Public key of account on Polkadot-like chains.
+pub type AccountPublic = <Signature as Verify>::Signer;
+
+/// Id of account on Polkadot-like chains.
+pub type AccountId = <AccountPublic as IdentifyAccount>::AccountId;
+
+/// Index of a transaction on the Polkadot-like chains.
+pub type Nonce = u32;
+
+/// Block type of Polkadot-like chains.
+pub type Block = generic::Block<Header, UncheckedExtrinsic>;
+
+/// Polkadot-like block signed with a Justification.
+pub type SignedBlock = generic::SignedBlock<Block>;
+
+/// The balance of an account on Polkadot-like chain.
+pub type Balance = u128;
+
+/// Polkadot-like chain.
+#[derive(RuntimeDebug)]
+pub struct PolkadotLike;
+
+impl Chain for PolkadotLike {
+	type BlockNumber = BlockNumber;
+	type Hash = Hash;
+	type Hasher = Hasher;
+	type Header = Header;
+}
+
+/// Convert a 256-bit hash into an AccountId.
+pub struct AccountIdConverter;
+
+impl sp_runtime::traits::Convert<sp_core::H256, AccountId> for AccountIdConverter {
+	fn convert(hash: sp_core::H256) -> AccountId {
+		hash.to_fixed_bytes().into()
+	}
+}
+
+/*
+ * TODO [ToDr] Generate via macro?
 /// Name of the `PolkadotHeaderApi::best_blocks` runtime method.
 pub const BEST_POLKADOT_BLOCKS_METHOD: &str = "PolkadotHeaderApi_best_blocks";
 /// Name of the `PolkadotHeaderApi::finalized_block` runtime method.
@@ -94,3 +159,4 @@ sp_api::decl_runtime_apis! {
 		fn latest_confirmed_nonce(lane: LaneId) -> MessageNonce;
 	}
 }
+*/
